refactor(package): add package to apply MVC pattern
