+++ src/main/java/racingcar/MainController.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.domain.Game;
+
 public class MainController {
 
     public static void main(String[] args) {
