+++ src/test/java/racingcar/InputValidatorTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 import org.junit.jupiter.api.Test;
+import racingcar.view.InputValidator;
 
 @SuppressWarnings("NonAsciiCharacters")
 public class InputValidatorTest {
